From dev to main (week #2)

+++ src/main/java/ru/skypro/homework/dto/comment/Comment.java
@@ -4,10 +4,10 @@ import lombok.Data;
 
 @Data
 public class Comment {
-    private long author;
+    private int author;
     private String authorImage;
     private String authorFirstName;
     private long createdAt;
-    private long pk;
+    private int pk;
     private String text;
 }
